refactor: UpdateViewState
